+++ supvisors/tests/test_viewhandler.py
@@ -180,8 +180,19 @@ def test_write_nav(mocker, handler):
     """ Test the write_nav method. """
     mocked_appli = mocker.patch('supvisors.web.viewhandler.ViewHandler.write_nav_applications')
     mocked_instances = mocker.patch('supvisors.web.viewhandler.ViewHandler.write_nav_instances')
+    # test identifier set only
+    handler.write_nav('root', identifier='identifier')
+    assert mocked_instances.call_args_list == [call('root', 'identifier', False)]
+    assert mocked_appli.call_args_list == [call('root', None)]
+    mocker.resetall()
+    # test application set only
+    handler.write_nav('root', appli='appli')
+    assert mocked_instances.call_args_list == [call('root', None, True)]
+    assert mocked_appli.call_args_list == [call('root', 'appli')]
+    mocker.resetall()
+    # test both set
     handler.write_nav('root', 'identifier', 'appli')
-    assert mocked_instances.call_args_list == [call('root', 'identifier')]
+    assert mocked_instances.call_args_list == [call('root', 'identifier', True)]
     assert mocked_appli.call_args_list == [call('root', 'appli')]
 
 
@@ -195,7 +206,7 @@ def test_write_nav_instances_identifier_error(supvisors, handler):
     mocked_mid = Mock(**{'repeat.return_value': [(address_elt, '10.0.0.0')]})
     mocked_root = Mock(**{'findmeld.return_value': mocked_mid})
     # test call with no address status in context
-    handler.write_nav_instances(mocked_root, '10.0.0.0')
+    handler.write_nav_instances(mocked_root, '10.0.0.0', False)
     assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
     fqdn = socket.getfqdn()
     expected = ['10.0.0.1:25000', '10.0.0.2:25000', '10.0.0.3:25000', '10.0.0.4:25000', '10.0.0.5:25000',
@@ -217,7 +228,7 @@ def test_write_nav_instances_silent_instance(supvisors, handler):
     mocked_root = create_element({'instance_h_mid': instance_h_mid, 'instance_li_mid': instance_li_mid})
     # test call with address status set in context, SILENT and different from parameter
     handler.sup_ctx.instances['10.0.0.1:25000']._state = SupvisorsInstanceStates.SILENT
-    handler.write_nav_instances(mocked_root, '10.0.0.2:25000')
+    handler.write_nav_instances(mocked_root, '10.0.0.2:25000', False)
     assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
     assert instance_li_mid.repeat.call_args_list == [call(list(supvisors.mapper.instances.keys()))]
     assert instance_elt.attrib['class'] == 'SILENT'
@@ -227,11 +238,22 @@ def test_write_nav_instances_silent_instance(supvisors, handler):
     assert not master_sp_mid.content.called
     instance_elt.reset_all()
     mocked_root.reset_all()
-    # test call with address status set in context, SILENT and identical to parameter
-    handler.write_nav_instances(mocked_root, '10.0.0.1:25000')
+    # test call with address status set in context, SILENT, identical to parameter and not marked as location
+    handler.write_nav_instances(mocked_root, '10.0.0.1:25000', False)
     assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
     assert instance_li_mid.repeat.call_args_list == [call(list(supvisors.mapper.instances.keys()))]
     assert instance_elt.attrib['class'] == 'SILENT active'
+    assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
+    assert instance_a_mid.attrib['class'] == 'off'
+    assert instance_sp_mid.content.call_args_list == [call('10.0.0.1')]
+    assert not master_sp_mid.content.called
+    instance_elt.reset_all()
+    mocked_root.reset_all()
+    # test call with address status set in context, SILENT, identical to parameter and not marked as location
+    handler.write_nav_instances(mocked_root, '10.0.0.1:25000', True)
+    assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
+    assert instance_li_mid.repeat.call_args_list == [call(list(supvisors.mapper.instances.keys()))]
+    assert instance_elt.attrib['class'] == 'SILENT local'
     assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
     assert instance_a_mid.attrib['class'] == 'off'
     assert instance_sp_mid.content.call_args_list == [call('10.0.0.1')]
@@ -257,10 +279,11 @@ def test_write_nav_instances_running_instance(mocker, supvisors, handler):
         # set context
         status._state = state
         status.state_modes.starting_jobs = False
+        status.state_modes.stopping_jobs = False
         handler.sup_ctx.master_identifier = ''
-        # test call with address status set in context, RUNNING, different from parameter and not MASTER
+        # test call with address status set in context, different from parameter and not MASTER
         mocker.patch.object(status, 'has_error', return_value=False)
-        handler.write_nav_instances(mocked_root, '10.0.0.2:25000')
+        handler.write_nav_instances(mocked_root, '10.0.0.2:25000', True)
         assert mocked_root.findmeld.call_args_list == [call('instance_li_mid')]
         assert instance_li_mid.repeat.call_args_list == [call(all_identifiers)]
         assert instance_elt.attrib['class'] == state.name
@@ -270,20 +293,43 @@ def test_write_nav_instances_running_instance(mocker, supvisors, handler):
         assert instance_a_mid.attrib['class'] == 'on'
         assert instance_sp_mid.content.call_args_list == [call('10.0.0.1')]
         assert not master_sp_mid.content.called
+        mocker.resetall()
         instance_elt.reset_all()
         mocked_root.reset_all()
         handler.view_ctx.format_url.reset_mock()
-        # test call with address status set in context, RUNNING, identical to parameter and MASTER
-        # failure is added to the instance
+        # test call with address status set in context, identical to parameter, not marked as location
+        # and not MASTER
+        status.state_modes.starting_jobs = False
+        status.state_modes.stopping_jobs = True
+        handler.sup_ctx.master_identifier = '10.0.0.1:25000'
+        handler.write_nav_instances(mocked_root, '10.0.0.1:25000', False)
+        expected = [call('instance_li_mid')]
+        assert mocked_root.findmeld.call_args_list == expected
+        assert instance_li_mid.repeat.call_args_list == [call(all_identifiers)]
+        assert instance_elt.attrib['class'] == state.name + ' active master'
+        assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
+        assert handler.view_ctx.format_url.call_args_list == [call('10.0.0.1:25000', 'proc_instance.html')]
+        assert instance_a_mid.attributes.call_args_list == [call(href='an url')]
+        assert instance_a_mid.attrib['class'] == 'blink on'
+        assert instance_sp_mid.content.call_args_list == [call('10.0.0.1')]
+        assert master_sp_mid.content.call_args_list == [call(MASTER_SYMBOL)]
+        assert instance_h_mid.attrib['class'] == ''
+        mocker.resetall()
+        instance_elt.reset_all()
+        mocked_root.reset_all()
+        handler.view_ctx.format_url.reset_mock()
+        # test call with address status set in context, identical to parameter, marked as location
+        # and MASTER failure is added to the instance
         is_running = state == SupvisorsInstanceStates.RUNNING
         mocker.patch.object(status, 'has_error', return_value=is_running)
         status.state_modes.starting_jobs = True
+        status.state_modes.stopping_jobs = False
         handler.sup_ctx.master_identifier = '10.0.0.1:25000'
-        handler.write_nav_instances(mocked_root, '10.0.0.1:25000')
+        handler.write_nav_instances(mocked_root, '10.0.0.1:25000', True)
         expected = [call('instance_li_mid')] + ([call('instance_h_mid')] if is_running else [])
         assert mocked_root.findmeld.call_args_list == expected
         assert instance_li_mid.repeat.call_args_list == [call(all_identifiers)]
-        assert instance_elt.attrib['class'] == state.name + ' active' + (' failure' if is_running else '') + ' master'
+        assert instance_elt.attrib['class'] == state.name + ' local' + (' failure' if is_running else '') + ' master'
         assert instance_elt.findmeld.call_args_list == [call('instance_a_mid')]
         assert handler.view_ctx.format_url.call_args_list == [call('10.0.0.1:25000', 'proc_instance.html')]
         assert instance_a_mid.attributes.call_args_list == [call(href='an url')]
@@ -291,6 +337,7 @@ def test_write_nav_instances_running_instance(mocker, supvisors, handler):
         assert instance_sp_mid.content.call_args_list == [call('10.0.0.1')]
         assert master_sp_mid.content.call_args_list == [call(MASTER_SYMBOL)]
         assert instance_h_mid.attrib['class'] == ('failure' if is_running else '')
+        mocker.resetall()
         instance_elt.reset_all()
         mocked_root.reset_all()
         handler.view_ctx.format_url.reset_mock()